movingFundsTimeoutResetDelay uint64->uint32

+++ solidity/contracts/bridge/BridgeGovernance.sol
@@ -138,7 +138,7 @@ contract BridgeGovernance is Ownable {
         uint256 timestamp
     );
     event MovingFundsTimeoutResetDelayUpdated(
-        uint64 movingFundsTimeoutResetDelay
+        uint32 movingFundsTimeoutResetDelay
     );
 
     event MovingFundsTimeoutUpdateStarted(
